refactor(models): split models.py into package; replace parse_X with from_congress_api
Splits the 1130-line src/concord/models.py into a package of focused
sub-modules (_common, proceedings, members, bills, votes), each well
under 400 lines. Public imports (`from concord.models import …`) are
unchanged.

Replaces the `parse_X(payload: dict) -> X | None` paradigm with
`X.from_congress_api(payload) -> X` classmethods that raise
`ValueError` / `pydantic.ValidationError` on contract violations
instead of silently returning None. Loaders now catch the failure and
log the offending payload, surfacing data-contract issues we used to
drop on the floor.

Also inlines two single-use Literal aliases (`BillType`,
`OriginChamber`) into `Bill` and drops the unused `BILL_TYPES` tuple.
