+++ src/concord/models/members.py
@@ -1,0 +1,323 @@
+"""Member, Term, and MemberSnapshot models (Phase 1).
+
+The list endpoint ``/v3/member/congress/{n}`` returns the **same** payload
+regardless of which Congress was queried, so the projection from payload
+to :class:`Term` always requires the queried-Congress context.
+"""
+
+from datetime import datetime
+from typing import Any
+
+from pydantic import BaseModel, ConfigDict, field_validator
+
+from concord.models._common import Chamber, coerce_int, normalize_chamber
+
+_STATE_NAME_TO_CODE = {
+    "alabama": "AL",
+    "alaska": "AK",
+    "arizona": "AZ",
+    "arkansas": "AR",
+    "california": "CA",
+    "colorado": "CO",
+    "connecticut": "CT",
+    "delaware": "DE",
+    "florida": "FL",
+    "georgia": "GA",
+    "hawaii": "HI",
+    "idaho": "ID",
+    "illinois": "IL",
+    "indiana": "IN",
+    "iowa": "IA",
+    "kansas": "KS",
+    "kentucky": "KY",
+    "louisiana": "LA",
+    "maine": "ME",
+    "maryland": "MD",
+    "massachusetts": "MA",
+    "michigan": "MI",
+    "minnesota": "MN",
+    "mississippi": "MS",
+    "missouri": "MO",
+    "montana": "MT",
+    "nebraska": "NE",
+    "nevada": "NV",
+    "new hampshire": "NH",
+    "new jersey": "NJ",
+    "new mexico": "NM",
+    "new york": "NY",
+    "north carolina": "NC",
+    "north dakota": "ND",
+    "ohio": "OH",
+    "oklahoma": "OK",
+    "oregon": "OR",
+    "pennsylvania": "PA",
+    "rhode island": "RI",
+    "south carolina": "SC",
+    "south dakota": "SD",
+    "tennessee": "TN",
+    "texas": "TX",
+    "utah": "UT",
+    "vermont": "VT",
+    "virginia": "VA",
+    "washington": "WA",
+    "west virginia": "WV",
+    "wisconsin": "WI",
+    "wyoming": "WY",
+    "district of columbia": "DC",
+    "puerto rico": "PR",
+    "american samoa": "AS",
+    "guam": "GU",
+    "northern mariana islands": "MP",
+    "virgin islands": "VI",
+}
+
+_STATE_CODE_LEN = 2
+
+
+def normalize_state(value: str | None) -> str | None:
+    """Map the API's full state name to a two-letter code; pass through codes."""
+    if value is None:
+        return None
+    stripped = value.strip()
+    if len(stripped) == _STATE_CODE_LEN and stripped.isupper():
+        return stripped
+    return _STATE_NAME_TO_CODE.get(stripped.lower(), stripped)
+
+
+class Term(BaseModel):
+    """One continuous service period for a :class:`Member` in one chamber.
+
+    Keyed by ``(bioguide_id, congress, chamber)``. Party, state, and district
+    are recorded per-Term so a Member who changed party between Congresses or
+    moved between House and Senate has multiple Terms reflecting reality.
+    """
+
+    model_config = ConfigDict(populate_by_name=True, extra="ignore")
+
+    bioguide_id: str
+    congress: int
+    chamber: Chamber
+    party: str | None = None
+    state: str  # two-letter state code (e.g. "VT")
+    district: int | None = None  # NULL for senators
+    start_date: str | None = None  # ISO YYYY-MM-DD or YYYY (year-only fallback)
+    end_date: str | None = None  # None = currently serving
+
+    @field_validator("chamber", mode="before")
+    @classmethod
+    def _coerce_chamber(cls, value: Any) -> Any:
+        return normalize_chamber(value)
+
+    @field_validator("state", mode="before")
+    @classmethod
+    def _coerce_state(cls, value: Any) -> Any:
+        return normalize_state(value) if isinstance(value, str) else value
+
+    @classmethod
+    def from_congress_api(cls, payload: dict[str, Any], *, congress: int) -> "Term":  # noqa: C901 — one branch per optional payload field
+        """Project a raw ``/member`` payload + queried Congress into one :class:`Term`.
+
+        The list endpoint omits ``congress`` from each ``terms.item`` and
+        returns the **same** payload for every Congress a Member served in;
+        the queried ``congress`` argument supplies that missing context.
+
+        Raises ``ValueError`` if no ``terms.item`` covers ``congress`` (the
+        payload contradicts its own listing) or if the matched item lacks a
+        usable chamber/state. Raises ``ValidationError`` if the projected
+        row fails Pydantic validation.
+        """
+        bioguide_id = str(payload["bioguideId"])
+        items = _extract_term_items(payload)
+        item = _term_item_for_congress(items, congress)
+        if item is None:
+            raise ValueError(
+                f"payload for {bioguide_id} has no terms.item covering Congress {congress}"
+            )
+
+        chamber_raw = item.get("chamber")
+        if chamber_raw is None:
+            raise ValueError(f"term item for {bioguide_id}/{congress} has no chamber")
+        chamber = normalize_chamber(chamber_raw)
+        if chamber not in {"house", "senate"}:
+            raise ValueError(
+                f"term item for {bioguide_id}/{congress} has unknown chamber {chamber_raw!r}"
+            )
+
+        top_state = payload.get("state")
+        top_district = payload.get("district")
+        top_party = payload.get("partyName")
+
+        state = item.get("stateCode") or item.get("stateName") or top_state
+        if not state:
+            raise ValueError(f"term item for {bioguide_id}/{congress} has no state")
+
+        district = item.get("district")
+        if district is None and chamber == "house":
+            district = top_district
+        if isinstance(district, str):
+            try:
+                district = int(district)
+            except ValueError:
+                district = None
+        if chamber == "senate":
+            district = None
+
+        # Clip the chamber-career window (API's startYear/endYear) to this
+        # Congress's window. Each (member, congress) row should describe the
+        # service inside that Congress, not the broader career stretch.
+        cong_start_year = 2 * congress + 1787  # Congress 1 began 1789-01-03
+        cong_end_year = 2 * congress + 1789  # = next Congress's start year
+
+        api_start = coerce_int(item.get("startYear"))
+        if api_start is not None and api_start > cong_start_year:
+            start_date = f"{api_start:04d}-01-03"  # joined mid-Congress
+        else:
+            start_date = f"{cong_start_year:04d}-01-03"
+
+        api_end = coerce_int(item.get("endYear"))
+        if api_end is not None and api_end < cong_end_year:
+            end_date = f"{api_end:04d}-01-03"  # left mid-Congress
+        else:
+            end_date = f"{cong_end_year:04d}-01-03"
+
+        return cls(
+            bioguide_id=bioguide_id,
+            congress=congress,
+            chamber=chamber,
+            party=item.get("partyName") or top_party,
+            state=state,  # validator normalizes to 2-letter
+            district=district,
+            start_date=start_date,
+            end_date=end_date,
+        )
+
+
+class Member(BaseModel):
+    """A person who has served in Congress.
+
+    Identity fields only — anything that varies across a career (party,
+    chamber, state, district) lives on :class:`Term` instead.
+    """
+
+    model_config = ConfigDict(populate_by_name=True, extra="ignore")
+
+    bioguide_id: str
+    first_name: str
+    middle_name: str | None = None
+    last_name: str
+    suffix: str | None = None
+    birth_year: int | None = None
+    death_year: int | None = None
+    display_name: str  # API's directOrderName
+    photo_url: str | None = None  # API's depiction.imageUrl
+    biography: str | None = None
+
+    @classmethod
+    def from_congress_api(cls, payload: dict[str, Any]) -> "Member":
+        """Project a raw ``/member`` payload into the identity-only Member row.
+
+        Identity fields (name, birth year, photo) are the same regardless of
+        which Congress the API was queried for, so this is the half of a
+        member payload that's safe to project without knowing the queried
+        Congress. Raises ``ValidationError`` if the projected row is malformed.
+        """
+        bioguide_id = str(payload["bioguideId"])
+
+        first_name = payload.get("firstName")
+        last_name = payload.get("lastName")
+        if not first_name or not last_name:
+            # Fall back to parsing the inverted "Last, First" name from the
+            # list endpoint when structured fields aren't present.
+            inverted = payload.get("invertedOrderName") or payload.get("name") or ""
+            derived_first, derived_last = _split_inverted_name(inverted)
+            first_name = first_name or derived_first
+            last_name = last_name or derived_last
+
+        display_name = payload.get("directOrderName") or (
+            f"{first_name} {last_name}".strip()
+            if first_name or last_name
+            else payload.get("name", "")
+        )
+
+        depiction = payload.get("depiction") or {}
+        photo_url = depiction.get("imageUrl") if isinstance(depiction, dict) else None
+
+        return cls(
+            bioguide_id=bioguide_id,
+            first_name=first_name or "",
+            middle_name=payload.get("middleName"),
+            last_name=last_name or "",
+            suffix=payload.get("suffixName"),
+            birth_year=coerce_int(payload.get("birthYear")),
+            death_year=coerce_int(payload.get("deathYear")),
+            display_name=display_name,
+            photo_url=photo_url,
+            biography=payload.get("biography"),
+        )
+
+
+class MemberSnapshot(BaseModel):
+    """ADR 0006 envelope wrapping one raw ``/member`` API response.
+
+    Each Stage 0 fetch appends one of these to ``data/members.jsonl``. The
+    Stage 1 loader groups by ``key["bioguide_id"]`` and keeps the latest
+    ``fetched_at`` per key.
+    """
+
+    model_config = ConfigDict(extra="ignore")
+
+    fetched_at: datetime
+    key: dict[str, str | int]
+    payload: dict[str, Any]
+
+
+def _split_inverted_name(name: str) -> tuple[str, str]:
+    """Parse the API's ``"Last, First[ Middle][, Suffix]"`` into ``(first, last)``.
+
+    A best-effort fallback when the API's structured first/last fields are
+    absent (the list endpoint sometimes only returns the inverted form).
+    """
+    parts = [p.strip() for p in name.split(",", 1)]
+    if len(parts) == 1:
+        return parts[0], ""
+    last = parts[0]
+    first = parts[1].split()[0] if parts[1] else ""
+    return first, last
+
+
+def _extract_term_items(payload: dict[str, Any]) -> list[dict[str, Any]]:
+    """Pull the ``terms.item`` list from a /member payload, tolerating both
+    list-endpoint (``{"terms": {"item": [...]}}``) and detail-endpoint
+    (``{"terms": [...]}``) shapes."""
+    terms_raw = payload.get("terms")
+    if isinstance(terms_raw, dict):
+        return [i for i in (terms_raw.get("item") or []) if isinstance(i, dict)]
+    if isinstance(terms_raw, list):
+        return [i for i in terms_raw if isinstance(i, dict)]
+    return []
+
+
+def _term_item_for_congress(items: list[dict[str, Any]], congress: int) -> dict[str, Any] | None:
+    """Return the term item whose year range contains ``congress``, or None.
+
+    A term item covers Congress N if its ``startYear`` falls on or before
+    N's last year, and its ``endYear`` (when set) falls on or after N's
+    first year. If multiple items overlap, returns the last match — useful
+    for the rare member who switches chambers mid-Congress (a member is
+    only listed under one chamber per Congress in practice).
+    """
+    cong_first_year = 2 * congress + 1787
+    cong_last_year = cong_first_year + 1
+    match: dict[str, Any] | None = None
+    for item in items:
+        start = coerce_int(item.get("startYear"))
+        if start is None or start > cong_last_year:
+            continue
+        end = coerce_int(item.get("endYear"))
+        if end is not None and end < cong_first_year:
+            continue
+        match = item
+    return match
+
+
+__all__ = ["Member", "MemberSnapshot", "Term", "normalize_state"]
