+++ tests/test_models_bills.py
@@ -9,14 +9,13 @@ from pydantic import ValidationError
 
 from concord.models import (
     Bill,
+    BillAction,
     BillSnapshot,
+    BillSubject,
+    BillSummary,
+    BillTitle,
+    Cosponsor,
     bill_id_from_components,
-    parse_bill,
-    parse_bill_action,
-    parse_bill_subject,
-    parse_bill_summary,
-    parse_bill_title,
-    parse_cosponsor,
 )
 
 from ._snapshots import wrap_snapshot
@@ -70,10 +69,10 @@ class TestBillModel:
             )
 
 
-class TestParseBill:
+class TestBillFromCongressApi:
     def test_parses_hr_1_fixture(self, fixtures_dir: Path) -> None:
         payload = json.loads((fixtures_dir / "api/bills/detail_119_hr_1.json").read_text())
-        bill = parse_bill(payload["bill"])
+        bill = Bill.from_congress_api(payload["bill"])
         assert bill.bill_id == "119-hr-1"
         assert bill.congress == 119
         assert bill.bill_type == "hr"
@@ -90,7 +89,7 @@ class TestParseBill:
 
     def test_parses_hr_22_fixture(self, fixtures_dir: Path) -> None:
         payload = json.loads((fixtures_dir / "api/bills/detail_119_hr_22.json").read_text())
-        bill = parse_bill(payload["bill"])
+        bill = Bill.from_congress_api(payload["bill"])
         assert bill.bill_id == "119-hr-22"
         assert bill.sponsor_bioguide_id == "R000614"
         assert bill.policy_area == "Government Operations and Politics"
@@ -104,19 +103,17 @@ class TestParseBill:
             "title": "Sponsorless Bill Act",
             "updateDate": "2026-04-01",
         }
-        bill = parse_bill(payload)
+        bill = Bill.from_congress_api(payload)
         assert bill.sponsor_bioguide_id is None
         assert bill.policy_area is None
         assert bill.latest_action_date is None
 
 
-class TestParseCosponsor:
+class TestCosponsorFromCongressApi:
     def test_parses_cosponsor_fixture(self, fixtures_dir: Path) -> None:
         payload = json.loads((fixtures_dir / "api/bills/cosponsors_119_hr_22.json").read_text())
-        rows = [parse_cosponsor(p) for p in payload["cosponsors"]]
-        assert all(r is not None for r in rows)
+        rows = [Cosponsor.from_congress_api(p) for p in payload["cosponsors"]]
         first = rows[0]
-        assert first is not None
         assert first.bioguide_id == "B001302"
         assert first.is_original_cosponsor is True
         assert first.sponsorship_withdrawn_date is None
@@ -125,59 +122,58 @@ class TestParseCosponsor:
         payload = json.loads(
             (fixtures_dir / "api/bills/cosponsors_119_hr_22_withdrawn.json").read_text()
         )
-        rows = [parse_cosponsor(p) for p in payload["cosponsors"]]
-        withdrawn = next(r for r in rows if r is not None and r.sponsorship_withdrawn_date)
+        rows = [Cosponsor.from_congress_api(p) for p in payload["cosponsors"]]
+        withdrawn = next(r for r in rows if r.sponsorship_withdrawn_date)
         assert withdrawn.sponsorship_withdrawn_date == "2025-04-02"
         assert withdrawn.is_original_cosponsor is False
 
-    def test_returns_none_for_row_without_bioguide(self) -> None:
-        assert parse_cosponsor({"sponsorshipDate": "2025-01-09"}) is None
+    def test_raises_for_row_without_bioguide(self) -> None:
+        with pytest.raises(ValueError, match="bioguideId"):
+            Cosponsor.from_congress_api({"sponsorshipDate": "2025-01-09"})
 
 
-class TestParseBillAction:
+class TestBillActionFromCongressApi:
     def test_parses_action_fixture(self, fixtures_dir: Path) -> None:
         payload = json.loads((fixtures_dir / "api/bills/actions_119_hr_1.json").read_text())
-        rows = [parse_bill_action(p) for p in payload["actions"]]
-        assert all(r is not None for r in rows)
+        rows = [BillAction.from_congress_api(p) for p in payload["actions"]]
         first = rows[0]
-        assert first is not None
         assert first.action_date == "2026-03-30"
         assert first.action_text == "Became Public Law No: 119-1."
         assert first.source_system == "Library of Congress"
 
-    def test_returns_none_for_action_without_date(self) -> None:
-        assert parse_bill_action({"text": "Some action"}) is None
+    def test_raises_for_action_without_date(self) -> None:
+        with pytest.raises(ValueError, match="actionDate"):
+            BillAction.from_congress_api({"text": "Some action"})
 
 
-class TestParseBillSubject:
+class TestBillSubjectFromCongressApi:
     def test_parses_subjects_fixture(self, fixtures_dir: Path) -> None:
         payload = json.loads((fixtures_dir / "api/bills/subjects_119_hr_1.json").read_text())
-        rows = [parse_bill_subject(p) for p in payload["subjects"]["legislativeSubjects"]]
-        assert all(r is not None for r in rows)
-        names = [r.name for r in rows if r is not None]
+        rows = [
+            BillSubject.from_congress_api(p) for p in payload["subjects"]["legislativeSubjects"]
+        ]
+        names = [r.name for r in rows]
         assert "Energy" in names
         assert "Pipelines" in names
 
-    def test_returns_none_for_empty_name(self) -> None:
-        assert parse_bill_subject({"name": ""}) is None
+    def test_raises_for_empty_name(self) -> None:
+        with pytest.raises(ValueError, match="name"):
+            BillSubject.from_congress_api({"name": ""})
 
 
-class TestParseBillTitle:
+class TestBillTitleFromCongressApi:
     def test_parses_titles_fixture(self, fixtures_dir: Path) -> None:
         payload = json.loads((fixtures_dir / "api/bills/titles_119_hr_1.json").read_text())
-        rows = [parse_bill_title(p) for p in payload["titles"]]
-        assert all(r is not None for r in rows)
-        short = next(r for r in rows if r is not None and r.title_type.startswith("Short Title"))
+        rows = [BillTitle.from_congress_api(p) for p in payload["titles"]]
+        short = next(r for r in rows if r.title_type.startswith("Short Title"))
         assert short.title_text == "Lower Energy Costs Act"
 
 
-class TestParseBillSummary:
+class TestBillSummaryFromCongressApi:
     def test_parses_summaries_fixture(self, fixtures_dir: Path) -> None:
         payload = json.loads((fixtures_dir / "api/bills/summaries_119_hr_1.json").read_text())
-        rows = [parse_bill_summary(p) for p in payload["summaries"]]
-        assert all(r is not None for r in rows)
+        rows = [BillSummary.from_congress_api(p) for p in payload["summaries"]]
         first = rows[0]
-        assert first is not None
         assert first.version_code == "00"
         assert "<p>" in first.summary_text
